backend/python-ai/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-multipart==0.0.20
redis==5.2.1
boto3==1.35.99
torch==2.5.1
torchvision==0.20.1
diffusers==0.32.2
transformers==4.47.1
accelerate==1.2.1
safetensors==0.4.5
opencv-python==4.10.0.84
pillow==11.0.0
numpy==2.2.1
realesrgan==0.3.0
basicsr==1.4.2
gfpgan==1.3.8
insightface==0.7.3
onnx==1.16.1
onnxruntime-gpu==1.18.1
pyngrok==7.2.2
bitsandbytes==0.45.0
peft==0.14.0
httpx==0.28.1
requests==2.32.3
scipy==1.15.1
